show photo

+++ src/components/AlbumList.js
@@ -14,6 +14,8 @@ export default class AlbumList extends Component {
     this.editAlbum=this.editAlbum.bind(this);
     this.saveMe=this.saveMe.bind(this);
     this.deleteAlbum=this.deleteAlbum.bind(this);
+    this.addPhotos=this.addPhotos.bind(this);
+    this.lookAtPhotos=this.lookAtPhotos.bind(this);
   }
 
   editAlbum(id){
@@ -34,7 +36,18 @@ export default class AlbumList extends Component {
   }
 
   deleteAlbum(id){
-    AlbumActions.deleteAlbum(id);
+    let r = confirm("Delete the Album. Are you sure?");
+    if (r == true) {
+      AlbumActions.deleteAlbum(id);
+    }
+  }
+
+  addPhotos(id){
+    browserHistory.push({pathname:'/albums/allPhotos', query:{ albumId : id}});
+  }
+
+  lookAtPhotos(){
+    browserHistory.push('/albums/intoAlbums');
   }
 
   render() {
@@ -60,17 +73,22 @@ export default class AlbumList extends Component {
           <tr className='trFont'>
             <td>{name}</td>
             <td>
-              <button id = {_id} className="btn btn-success btn-xs">
+              <button className="btn btn-success btn-xs" onClick={()=>this.addPhotos(_id)}>
                 <span className="glyphicon glyphicon-download-alt"></span>
               </button>
             </td>
             <td>
-              <button id = {_id} className="btn btn-primary btn-xs" onClick={()=>this.editAlbum(_id)}>
+              <button className="btn btn-success btn-xs" onClick={()=>this.lookAtPhotos()}>
+                <span className="glyphicon glyphicon-eye-open"></span>
+              </button>
+            </td>
+            <td>
+              <button className="btn btn-primary btn-xs" onClick={()=>this.editAlbum(_id)}>
                 <span className="glyphicon glyphicon-pencil"></span>
               </button>
             </td>
             <td>
-              <button id = {_id} className="btn btn-danger btn-xs" onClick={()=>this.deleteAlbum(_id)}>
+              <button className="btn btn-danger btn-xs" onClick={()=>this.deleteAlbum(_id)}>
                 <span className="glyphicon glyphicon-remove"></span>
               </button>
             </td>
